backend/requirements.txt
Flask==2.3.3
scikit-learn==1.4.1
pandas==2.2.2
numpy==1.26.4
requests==2.31.0
joblib==1.4.0
python-dotenv==1.0.1
matplotlib==3.8.4

